@@ -1,5 +1,5 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :consumer_complaint_company do
-    name "MyString"
+    name { Faker::Company.name }
   end
 end
