@@ -1,10 +1,8 @@
 module V1
   class ApplicationController < ActionController::API
 
-    def bad_request(message)
-      # @error = { error: 'bad request', message: message }
-      # render 'v1/events/error', format: :json, status: :bad_request
-    end
-
+    include FastJsonApiResponse
+    include ExceptionHandler
+    include Paginatable
   end
 end
