@@ -1,3 +1,4 @@
-class ConsumerComplaintCompanySerializer < ActiveModel::Serializer
+class ConsumerComplaintCompanySerializer
+  include FastJsonapi::ObjectSerializer
   attributes :id, :name
 end
