@@ -1,3 +1,4 @@
-class ConsumerComplaintIssueSerializer < ActiveModel::Serializer
+class ConsumerComplaintIssueSerializer
+  include FastJsonapi::ObjectSerializer
   attributes :id, :name
 end
