@@ -1,3 +1,4 @@
-class ConsumerComplaintSubissueSerializer < ActiveModel::Serializer
+class ConsumerComplaintSubissueSerializer
+  include FastJsonapi::ObjectSerializer
   attributes :id, :name
 end
