@@ -1,3 +1,4 @@
-class ConsumerComplaintSubproductSerializer < ActiveModel::Serializer
+class ConsumerComplaintSubproductSerializer
+  include FastJsonapi::ObjectSerializer
   attributes :id, :name
 end
