@@ -1,3 +1,4 @@
-class ConsumerComplaintZipCodeSerializer < ActiveModel::Serializer
+class ConsumerComplaintZipCodeSerializer
+  include FastJsonapi::ObjectSerializer
   attributes :id, :name
 end
